Fix template loading process
CodeMate will find template java files under the project, then compare
them with the existing template class files in <project root>/.codemate/
compiled_templates. If there is any match class file and it is newer,
use it directly.

When scan single code, CodeMate will first find the potential project
that covers the code, and use the templates that are availible.